[MSGPACK](https://msgpack.org/index.html) is known to be a faster alternative to JSON while fulfilling all the same features. [Go's msgpack](https://pkg.go.dev/github.com/vmihailenco/msgpack) is written to be a drop in replacement for code that already uses json, so I dropped in the replacement.

Benchmarks

```
# Old JSON

goos: linux
goarch: amd64
pkg: github.com/nakabonne/tstorage
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkStorage_InsertRows-8                  	 9356394	       432.3 ns/op	     229 B/op	       4 allocs/op
BenchmarkStorage_SelectAmongThousandPoints-8   	14496987	       369.8 ns/op	      56 B/op	       2 allocs/op
BenchmarkStorage_SelectAmongMillionPoints-8    	16722625	       268.8 ns/op	      56 B/op	       2 allocs/op
PASS
ok  	github.com/nakabonne/tstorage	17.837s
```

```
# New MSGPACK

goos: linux
goarch: amd64
pkg: github.com/nakabonne/tstorage
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkStorage_InsertRows-8                  	11978878	       409.6 ns/op	     228 B/op	       4 allocs/op
BenchmarkStorage_SelectAmongThousandPoints-8   	13840152	       352.6 ns/op	      56 B/op	       2 allocs/op
BenchmarkStorage_SelectAmongMillionPoints-8    	17263482	       268.4 ns/op	      56 B/op	       2 allocs/op
PASS
ok  	github.com/nakabonne/tstorage	22.404s
```

As we can see there's a 22.6ns performance improvement in insertions representing a 5% performance buff. This aswell as smaller encoded bits yield smaller DBs and less ram usage.